match service added

+++ src/main/java/com/ganeshtakale/ipldemo/repository/PlayersRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlayersRepository extends CrudRepository<Players,Integer>{
+public interface PlayersRepository extends CrudRepository<Players,Long>{
 
 	Iterable<Players> findByTeam_Id(long teamId);
